% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_colors.R
\name{add_data_colors}
\alias{add_data_colors}
\title{Add color to data columns in a data frame}
\usage{
add_data_colors(
  data,
  palette_colors = NULL,
  palette_type = "qualitative",
  palette = NULL,
  color_by = NULL,
  na_color = NULL,
  na_label = "(NA)"
)
}
\arguments{
\item{data}{A data frame to which colors will be added}

\item{palette_colors}{A named vector of color codes or a data frame with color codes and categories}

\item{palette_type}{A string specifying the name of the color palette to use}

\item{palette}{A vector of color codes to use as the color palette}

\item{color_by}{A string specifying the name of the column to use for coloring}
}
\value{
A copy of the input data frame with colors added to the specified data column
}
\description{
This function adds color to the data columns in a data frame or fringe,
using a color palette provided by the user.
}
\examples{
# Create some sample data
data <- data.frame(x = 1:5, y = c("A", "B", "C", "D", "E"))

# Define the colors for the data columns
colors <- c("red", "green", "blue", "orange", "purple")

# Apply the colors to the data
data_with_colors <- add_data_colors(data, colors, color_by = "y")

# View the result
print(data_with_colors)
}
