% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_labels.R
\name{add_data_tooltip}
\alias{add_data_tooltip}
\title{Add tooltips to data columns in a data frame}
\usage{
add_data_tooltip(data, nms = NULL, tooltip = NULL, dic = NULL)
}
\arguments{
\item{data}{A data frame to which tooltips will be added}

\item{nms}{A named character vector containing the labels for the data columns}

\item{tooltip}{A tooltip template}

\item{dic}{optional dic}
}
\value{
A copy of the input data frame with tooltips added to each data column
}
\description{
This function adds tooltips to the data columns in a data frame, using labels
provided by the user.
}
\examples{
# Create some sample data
data <- data.frame(x = 1:5, y = c("A", "B", "C", "D", "E"))

# Define the labels for the data columns
nms <- create_nms(data_names = names(data))

# Apply the tooltip to the data
data_with_tooltip <- add_data_tooltip(data, nms)

# View the result
print(data_with_tooltip)
}
