% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_arrange.R
\name{add_group_index}
\alias{add_group_index}
\title{Add group index to a data frame based on specified group columns}
\usage{
add_group_index(data, group_cols, index_cols = NULL)
}
\description{
Add group index to a data frame based on specified group columns
}
\keyword{internal}
