% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_labels.R
\name{create_nms}
\alias{create_nms}
\title{Create list of names from input data or dictionary}
\usage{
create_nms(data_names = NULL, dic = NULL)
}
\arguments{
\item{data_names}{A character vector of data names}

\item{dic}{A data frame containing variable IDs and their corresponding names}
}
\value{
A list of names
}
\description{
This function creates a list of names based on the input data or dictionary.
If a dictionary is provided, it is used to map variable IDs to their
corresponding names. Otherwise, the data names are used as the names of the list.
}
\examples{
data_names <- c("x", "y", "z")
dic <- data.frame(id = c("x", "y", "z"), label = c("Variable X", "Variable Y", "Variable Z"))
create_nms(data_names, dic)

}
