% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_data_filter.R
\name{data_filter}
\alias{data_filter}
\title{Data Filtering Function}
\usage{
data_filter(data, dic, var_inputs, special_placeholder = NULL, .id = NULL)
}
\arguments{
\item{data}{A data.frame that will be filtered.}

\item{dic}{A dictionary data.frame containing the 'id' of variable names and their data types.}

\item{var_inputs}{A list of variables to filter, where names are variable names and values are the desired values or ranges.}

\item{special_placeholder}{(Optional) A placeholder value to indicate special conditions.}

\item{.id}{(Optional) A special identifier for dealing with list variables.}
}
\value{
A filtered data.frame.
}
\description{
This function filters a dataset based on input specifications.
}
