% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdataprep-package.R
\docType{package}
\name{dsdataprep-package}
\alias{dsdataprep}
\alias{dsdataprep-package}
\title{dsdataprep: Data preparation}
\description{
dsdataprep helps manipulationdata by offering functions to aggregate, summarise, sort and wrap data; adds color to the data columns and adds rows information for displaying in interactive plots.
}
\author{
\strong{Maintainer}: Camila Achuri \email{camila@randommonkey.io}

}
\keyword{internal}
