% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_labels.R
\name{labels_map}
\alias{labels_map}
\title{Create tooltip string from list of labels}
\usage{
labels_map(nms, tooltip = NULL)
}
\arguments{
\item{nms}{A list of labels}

\item{tooltip}{A tooltip template}
}
\value{
A tooltip string
}
\description{
This function takes a list of labels and returns a tooltip string that can be
used to display the labels in a plot or other visualization.
Each label in the list is mapped to its name, and the
resulting strings are concatenated with line breaks ("\if{html}{\out{<br/>}}") between them.
}
\examples{
nms <- list(x = "Variable X", y = "Variable Y", z = "Variable Z")
labels_map(nms)

}
