% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_arrange.R
\name{numeric_sort}
\alias{numeric_sort}
\title{Sorts a data frame by a numeric column.}
\usage{
numeric_sort(
  data,
  col_num,
  col_cat = NULL,
  sort = NULL,
  slice_n = NULL,
  intra_cat = TRUE
)
}
\arguments{
\item{data}{a data frame to sort.}

\item{col_num}{the name of the numeric column to sort by.}

\item{col_cat}{the name of the categorical column(s) to group by before sorting.}

\item{sort}{the order to sort the data frame in (either "asc" or "desc").}

\item{slice_n}{an integer indicating the number of top rows to keep for each group}

\item{intra_cat}{a boolean indicating whether to slice within each category or across all categories}
}
\value{
the sorted data frame.
}
\description{
This function sorts a data frame by a numeric column. It can also group by
one or more categorical variables before sorting.
}
\keyword{internal}
