% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tooltip.R
\name{prep_tooltip}
\alias{prep_tooltip}
\title{Prepare tooltips for data in a table or plot}
\usage{
prep_tooltip(
  data,
  tooltip = NULL,
  new_labels = NULL,
  engine = "html",
  as_df = FALSE,
  na_row_default_column = NULL,
  na_row_default_value = NULL,
  na_label = "(NA)",
  format_num = NULL,
  opts_format_num = NULL,
  format_cat = NULL,
  format_date = NULL
)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{tooltip}{An optional tooltip template string that specifies how to format the data for display. If not specified, a default template will be created based on the column names of the data frame.}

\item{new_labels}{An optional named character vector of column labels to use in the tooltip. If specified, the column labels in the tooltip will be replaced with the new labels.}

\item{engine}{A character string specifying the templating engine to use. Currently supported options include "html", "markdown" and "latex".}

\item{as_df}{A logical value indicating whether to return the tooltip strings as a new column in the input data frame. If \code{FALSE} (the default), the tooltip strings will be returned as a character vector.}

\item{na_row_default_column}{An optional character string specifying the name of a column to use as the default value when all other columns in a row are \code{NA}. If specified, the values in this column will be used as the default tooltip value. If not specified, \code{na_row_default_value} will be used instead.}

\item{na_row_default_value}{An optional character string to use as the default tooltip value when all columns in a row are \code{NA}. If \code{NULL} (the default), the string "(NA)" will be used.}

\item{na_label}{An optional character string to use for \code{NA} values in the tooltip. If not specified, the string "(NA)" will be used.}

\item{format_num}{A named list of arguments to pass to the \code{\link{makeup_num}} function.}

\item{opts_format_num}{A named list of options to use for formatting numeric data.}

\item{format_cat}{A named list of arguments to pass to the \code{\link{makeup_chr}} function.}

\item{format_date}{A named list of arguments to pass to the \code{\link{makeup_dat}} function.}
}
\value{
A character vector of tooltip strings, or a data frame with a new column of tooltip strings.
}
\description{
This function generates a tooltip string for each row of a data frame or matrix. Tooltips are typically used to display additional information about the data when the user hovers over a data point in a plot or table.
}
