% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_arrange.R
\name{wrap_sort_data}
\alias{wrap_sort_data}
\title{Sort and wrap data by a categorical variable and numeric variable}
\usage{
wrap_sort_data(
  data,
  col_cat = NULL,
  col_num = NULL,
  order = NULL,
  order_legend = NULL,
  label_wrap = NULL,
  label_wrap_legend = NULL,
  new_line = "<br/>",
  sort = NULL,
  sort_by_cat = FALSE,
  slice_n = NULL,
  intra_cat = TRUE,
  index_names = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{col_cat}{A character vector specifying the name of the categorical variable to sort and wrap by.}

\item{col_num}{A character vector specifying the name of the numeric variable to sort by.}

\item{order}{A vector specifying the order of the categories.}

\item{order_legend}{A vector specifying the order of the categories in legend.}

\item{label_wrap}{An integer specifying the width of the wrapped label.}

\item{label_wrap_legend}{An integer specifying the width of the wrapped legend items.}

\item{new_line}{A string that will be used to replace newline characters.}

\item{sort}{A character vector that specifies the sorting order.}

\item{slice_n}{an integer indicating the number of top rows to keep for each group}

\item{intra_cat}{a boolean indicating whether to slice within each category or across all categories}

\item{index_names}{the name(s) of the new(s) index column(s)}
}
\value{
A data frame sorted and/or wrapped by col_cat and/or col_num.
}
\description{
Sort and wrap data by a categorical variable and numeric variable
}
